pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "work24-mcp-server"
version = "0.1.0"
description = "Work24 (고용24) OPEN API MCP Server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "xmltodict>=0.13.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
work24-mcp = "server:main"

[tool.setuptools]
packages = ["models", "tools", "utils"]
py-modules = ["server"]
